// src/app/Model/DistributorsModel.ts
export class distributorsDetailsModel {
    condition: string;
    firm_name: string
    name: string;
    mobile: string;
    address: string;
     branch:string;
}
export class retailerDetailsModel{
    condition: string;
    firm_name: string
    name: string;
    gst: string;
    contact_no: string;
    address: string;
    branch:string;
    grade: string;
    distributor:string;
}